# packages/api-client/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/sources":{"get":{"tags":["source-controller"],"operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SourcePageResponse"}}}}}},"post":{"tags":["source-controller"],"operationId":"upload","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}},"application/json":{"schema":{"$ref":"#/components/schemas/ImportSourceRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SourceResponse"}}}}}}},"/api/v1/sources/{id}/ocr":{"post":{"tags":["source-controller"],"operationId":"triggerOcr","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OcrTriggerResponse"}}}}}}},"/api/v1/sessions":{"post":{"tags":["session-controller"],"operationId":"start","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/api/v1/sessions/{id}/skip":{"post":{"tags":["session-controller"],"operationId":"skip","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkipRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SkipResponse"}}}}}}},"/api/v1/sessions/{id}/finish":{"post":{"tags":["session-controller"],"operationId":"finish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FinishSessionResponse"}}}}}}},"/api/v1/sessions/{id}/attempts":{"post":{"tags":["session-controller"],"operationId":"submitAttempt","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitAttemptRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SubmitAttemptResponse"}}}}}}},"/api/v1/items:bulk-reject":{"post":{"tags":["item-controller"],"operationId":"bulkReject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkItemIdsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"}}}}}}}},"/api/v1/items:bulk-publish":{"post":{"tags":["item-controller"],"operationId":"bulkPublish","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkItemIdsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"}}}}}}}},"/api/v1/items/{id}/report":{"post":{"tags":["item-controller"],"operationId":"report","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportItemRequest"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/items/{id}/reject":{"post":{"tags":["item-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemResponse"}}}}}}},"/api/v1/items/{id}/publish":{"post":{"tags":["item-controller"],"operationId":"publish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemResponse"}}}}}}},"/api/v1/concepts/{conceptId}/items:generate":{"post":{"tags":["item-controller"],"operationId":"generate","parameters":[{"name":"conceptId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateItemsRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenerateItemsResponse"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["auth-controller"],"operationId":"ssoToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SsoTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/ai/credentials":{"get":{"tags":["ai-credential-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CredentialResponse"}}}}}}},"post":{"tags":["ai-credential-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCredentialRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CredentialResponse"}}}}}}},"/api/v1/ai/credentials/{id}/verify":{"post":{"tags":["ai-credential-controller"],"operationId":"verify","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerifyCredentialResponse"}}}}}}},"/api/v1/sources/{id}/sections/{sectionId}":{"patch":{"tags":["source-controller"],"operationId":"excludeSection","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"sectionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExcludeSectionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SectionResponse"}}}}}}},"/api/v1/sources/{sourceId}/concepts":{"get":{"tags":["concept-controller"],"operationId":"list_2","parameters":[{"name":"sourceId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConceptResponse"}}}}}}}},"/api/v1/sources/{id}":{"get":{"tags":["source-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SourceResponse"}}}}}}},"/api/v1/sources/{id}/sections":{"get":{"tags":["source-controller"],"operationId":"sections","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SectionResponse"}}}}}}}},"/api/v1/sources/{id}/events":{"get":{"tags":["source-controller"],"operationId":"events","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/v1/sessions/{id}":{"get":{"tags":["session-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/api/v1/sessions/{id}/next":{"get":{"tags":["session-controller"],"operationId":"next","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NextItemResponse"}}}}}}},"/api/v1/reports/weekly":{"get":{"tags":["analytics-controller"],"operationId":"weeklyReport","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeeklyReportResponse"}}}}}}},"/api/v1/progress/overview":{"get":{"tags":["progress-controller"],"operationId":"overview","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProgressOverviewResponse"}}}}}}},"/api/v1/progress/misconceptions":{"get":{"tags":["analytics-controller"],"operationId":"misconceptions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MisconceptionResponse"}}}}}}}},"/api/v1/progress/due":{"get":{"tags":["progress-controller"],"operationId":"due","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConceptProgressResponse"}}}}}}}},"/api/v1/progress/concepts":{"get":{"tags":["progress-controller"],"operationId":"concepts","parameters":[{"name":"sourceId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConceptProgressResponse"}}}}}}}},"/api/v1/me":{"get":{"tags":["me-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}},"delete":{"tags":["me-controller"],"operationId":"deleteMe","responses":{"204":{"description":"No Content"}}}},"/api/v1/items/review-queue":{"get":{"tags":["item-controller"],"operationId":"reviewQueue","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"DRAFT","enum":["DRAFT","PUBLISHED","REJECTED","RETIRED"]}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemPageResponse"}}}}}}},"/api/v1/concepts/{conceptId}/items":{"get":{"tags":["item-controller"],"operationId":"listForConcept","parameters":[{"name":"conceptId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"}}}}}}}},"/api/v1/ai/usage":{"get":{"tags":["ai-usage-controller"],"operationId":"usage","parameters":[{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsageSummaryResponse"}}}}}}},"/api/v1/ai/providers":{"get":{"tags":["ai-credential-controller"],"operationId":"providers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProviderInfo"}}}}}}}},"/api/v1/ai/credentials/{id}":{"delete":{"tags":["ai-credential-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"SourceResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"kind":{"type":"string"},"title":{"type":"string"},"status":{"type":"string"},"failureReason":{"type":"string"},"pageCount":{"type":"integer","format":"int32"},"needsOcr":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}},"required":["createdAt","id","kind","needsOcr","status","title"]},"ImportSourceRequest":{"type":"object","properties":{"url":{"type":"string"},"html":{"type":"string"},"title":{"type":"string"}}},"OcrTriggerResponse":{"type":"object","properties":{"jobId":{"type":"string","format":"uuid"}},"required":["jobId"]},"StartSessionRequest":{"type":"object","properties":{"scopeKind":{"type":"string"},"scopeId":{"type":"string","format":"uuid"},"goalKind":{"type":"string"},"goalValue":{"type":"integer","format":"int32"}},"required":["goalKind","goalValue","scopeKind"]},"ItemMeta":{"type":"object","properties":{"conceptId":{"type":"string","format":"uuid"},"expectedSuccess":{"type":"number","format":"float"}},"required":["conceptId","expectedSuccess"]},"JsonNode":{},"NextItemResponse":{"type":"object","properties":{"itemId":{"type":"string","format":"uuid"},"type":{"type":"string"},"stem":{"type":"string"},"payload":{"$ref":"#/components/schemas/JsonNode"},"meta":{"$ref":"#/components/schemas/ItemMeta"}},"required":["itemId","meta","payload","stem","type"]},"SessionResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"scopeKind":{"type":"string"},"scopeId":{"type":"string","format":"uuid"},"goalKind":{"type":"string"},"goalValue":{"type":"integer","format":"int32"},"startedAt":{"type":"string","format":"date-time"},"endedAt":{"type":"string","format":"date-time"},"next":{"$ref":"#/components/schemas/NextItemResponse"}},"required":["goalKind","goalValue","scopeKind","sessionId","startedAt"]},"SkipRequest":{"type":"object","properties":{"itemId":{"type":"string","format":"uuid"},"reason":{"type":"string"}},"required":["itemId"]},"SkipResponse":{"type":"object","properties":{"next":{"$ref":"#/components/schemas/NextItemResponse"}}},"FinishSessionResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"attemptCount":{"type":"integer","format":"int32"},"accuracy":{"type":"number","format":"float"},"durationMs":{"type":"integer","format":"int64"}},"required":["accuracy","attemptCount","durationMs","sessionId"]},"SubmitAttemptRequest":{"type":"object","properties":{"itemId":{"type":"string","format":"uuid"},"response":{"$ref":"#/components/schemas/JsonNode"},"elapsedMs":{"type":"integer","format":"int32"}},"required":["elapsedMs","itemId","response"]},"ErrorAnalysisResponse":{"type":"object","properties":{"category":{"type":"string"},"confidence":{"type":"number","format":"float"},"note":{"type":"string"}},"required":["category","confidence"]},"EvidenceResponse":{"type":"object","properties":{"quote":{"type":"string"},"sourceId":{"type":"string","format":"uuid"},"chunkId":{"type":"string","format":"uuid"},"page":{"type":"integer","format":"int32"}},"required":["chunkId","quote","sourceId"]},"FeedbackResponse":{"type":"object","properties":{"correctResponse":{"$ref":"#/components/schemas/JsonNode"},"explanation":{"type":"string"},"chosenOptionRationale":{"type":"string"},"evidence":{"$ref":"#/components/schemas/EvidenceResponse"}},"required":["correctResponse","explanation"]},"LearnerUpdateResponse":{"type":"object","properties":{"conceptId":{"type":"string","format":"uuid"},"thetaBefore":{"type":"number","format":"float"},"thetaAfter":{"type":"number","format":"float"},"mastery":{"type":"number","format":"float"},"nextDueAt":{"type":"string","format":"date-time"}},"required":["conceptId","mastery","nextDueAt","thetaAfter","thetaBefore"]},"SubmitAttemptResponse":{"type":"object","properties":{"attemptId":{"type":"integer","format":"int64"},"outcome":{"type":"string"},"score":{"type":"number","format":"float"},"feedback":{"$ref":"#/components/schemas/FeedbackResponse"},"errorAnalysis":{"$ref":"#/components/schemas/ErrorAnalysisResponse"},"learnerUpdate":{"$ref":"#/components/schemas/LearnerUpdateResponse"},"next":{"$ref":"#/components/schemas/NextItemResponse"}},"required":["attemptId","feedback","learnerUpdate","outcome","score"]},"BulkItemIdsRequest":{"type":"object","properties":{"ids":{"type":"array","items":{"type":"string","format":"uuid"}}},"required":["ids"]},"ClozeBlank":{"type":"object","properties":{"accepted":{"type":"array","items":{"type":"string"}}},"required":["accepted"]},"ClozePayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"template":{"type":"string"},"blanks":{"type":"array","items":{"$ref":"#/components/schemas/ClozeBlank"}}}}],"required":["blanks","template"]},"ItemPayload":{},"ItemResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"conceptId":{"type":"string","format":"uuid"},"type":{"type":"string"},"stem":{"type":"string"},"payload":{"oneOf":[{"$ref":"#/components/schemas/ClozePayload"},{"$ref":"#/components/schemas/MatchingPayload"},{"$ref":"#/components/schemas/McMultiPayload"},{"$ref":"#/components/schemas/McSinglePayload"},{"$ref":"#/components/schemas/OrderingPayload"},{"$ref":"#/components/schemas/TrueFalsePayload"}]},"explanation":{"type":"string"},"bloomLevel":{"type":"string"},"status":{"type":"string"},"sourceChunkId":{"type":"string","format":"uuid"},"sourceCharFrom":{"type":"integer","format":"int32"},"sourceCharTo":{"type":"integer","format":"int32"},"reportCount":{"type":"integer","format":"int32"},"generatedBy":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}},"required":["bloomLevel","conceptId","createdAt","explanation","id","payload","reportCount","sourceCharFrom","sourceCharTo","sourceChunkId","status","stem","type"]},"MatchingItem":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"}},"required":["id","text"]},"MatchingPair":{"type":"object","properties":{"leftId":{"type":"string"},"rightId":{"type":"string"}},"required":["leftId","rightId"]},"MatchingPayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"left":{"type":"array","items":{"$ref":"#/components/schemas/MatchingItem"}},"right":{"type":"array","items":{"$ref":"#/components/schemas/MatchingItem"}},"pairs":{"type":"array","items":{"$ref":"#/components/schemas/MatchingPair"}},"distractorsRight":{"type":"array","items":{"$ref":"#/components/schemas/MatchingItem"}}}}],"required":["distractorsRight","left","pairs","right"]},"McMultiPayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"options":{"type":"array","items":{"$ref":"#/components/schemas/Option"}},"shuffle":{"type":"boolean"}}}],"required":["options","shuffle"]},"McSinglePayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"options":{"type":"array","items":{"$ref":"#/components/schemas/Option"}},"shuffle":{"type":"boolean"}}}],"required":["options","shuffle"]},"Option":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"},"correct":{"type":"boolean"},"rationale":{"type":"string"},"misconceptionCategory":{"type":"string"}},"required":["correct","id","rationale","text"]},"OrderingElement":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"}},"required":["id","text"]},"OrderingPayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"elements":{"type":"array","items":{"$ref":"#/components/schemas/OrderingElement"}},"correctOrder":{"type":"array","items":{"type":"string"}},"partialCredit":{"type":"boolean"}}}],"required":["correctOrder","elements","partialCredit"]},"TrueFalsePayload":{"allOf":[{"$ref":"#/components/schemas/ItemPayload"},{"type":"object","properties":{"statement":{"type":"string"},"answer":{"type":"boolean"},"rationale":{"type":"string"}}}],"required":["answer","rationale","statement"]},"ReportItemRequest":{"type":"object","properties":{"reason":{"type":"string"},"comment":{"type":"string"}},"required":["reason"]},"GenerateItemsRequest":{"type":"object","properties":{"count":{"type":"integer","format":"int32"},"types":{"type":"array","items":{"type":"string"}}},"required":["count","types"]},"GenerateItemsResponse":{"type":"object","properties":{"jobId":{"type":"string","format":"uuid"}},"required":["jobId"]},"SsoTokenRequest":{"type":"object","properties":{"provider":{"type":"string","enum":["GOOGLE","GITHUB"]},"code":{"type":"string","minLength":1},"codeVerifier":{"type":"string","minLength":1},"redirectUri":{"type":"string","minLength":1}},"required":["code","codeVerifier","provider","redirectUri"]},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}},"required":["accessToken","expiresIn","refreshToken","tokenType"]},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":100,"minLength":8},"displayName":{"type":"string"}},"required":["email","password"]},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LogoutRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"CreateCredentialRequest":{"type":"object","properties":{"provider":{"type":"string","enum":["OPENAI","ANTHROPIC","GOOGLE","AZURE","OPENROUTER","OLLAMA"]},"apiKey":{"type":"string","minLength":1},"baseUrl":{"type":"string"},"label":{"type":"string"}},"required":["apiKey","provider"]},"CredentialResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"provider":{"type":"string","enum":["OPENAI","ANTHROPIC","GOOGLE","AZURE","OPENROUTER","OLLAMA"]},"label":{"type":"string"},"keyHint":{"type":"string"},"baseUrl":{"type":"string"},"region":{"type":"string"},"status":{"type":"string"},"lastVerifiedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}},"required":["createdAt","id","keyHint","provider","status"]},"VerifyCredentialResponse":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"lastVerifiedAt":{"type":"string","format":"date-time"}},"required":["message","status"]},"ExcludeSectionRequest":{"type":"object","properties":{"excluded":{"type":"boolean"}},"required":["excluded"]},"SectionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"parentId":{"type":"string","format":"uuid"},"ordinal":{"type":"integer","format":"int32"},"level":{"type":"integer","format":"int32"},"title":{"type":"string"},"excluded":{"type":"boolean"}},"required":["excluded","id","level","ordinal","title"]},"SourcePageResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/SourceResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"}},"required":["items","page","size","totalElements"]},"ConceptResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"summary":{"type":"string"},"frequency":{"type":"integer","format":"int32"}},"required":["frequency","id","name","summary"]},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"WeeklyReportGapResponse":{"type":"object","properties":{"conceptId":{"type":"string","format":"uuid"},"conceptName":{"type":"string"},"mastery":{"type":"number","format":"float"}},"required":["conceptId","conceptName","mastery"]},"WeeklyReportResponse":{"type":"object","properties":{"topGaps":{"type":"array","items":{"$ref":"#/components/schemas/WeeklyReportGapResponse"}},"recommendedFocus":{"$ref":"#/components/schemas/WeeklyReportGapResponse"}},"required":["topGaps"]},"ProgressOverviewResponse":{"type":"object","properties":{"conceptCount":{"type":"integer","format":"int32"},"averageMastery":{"type":"number","format":"float"},"dueCount":{"type":"integer","format":"int32"}},"required":["averageMastery","conceptCount","dueCount"]},"MisconceptionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"conceptId":{"type":"string","format":"uuid"},"category":{"type":"string"},"occurrences":{"type":"integer","format":"int32"},"flagged":{"type":"boolean"},"firstSeenAt":{"type":"string","format":"date-time"},"lastSeenAt":{"type":"string","format":"date-time"}},"required":["category","conceptId","firstSeenAt","flagged","id","lastSeenAt","occurrences"]},"ConceptProgressResponse":{"type":"object","properties":{"conceptId":{"type":"string","format":"uuid"},"theta":{"type":"number","format":"float"},"mastery":{"type":"number","format":"float"},"state":{"type":"string"},"reps":{"type":"integer","format":"int32"},"lapses":{"type":"integer","format":"int32"},"dueAt":{"type":"string","format":"date-time"}},"required":["conceptId","lapses","mastery","reps","state","theta"]},"MeResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"locale":{"type":"string"},"plan":{"type":"string"},"workspaceId":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"}},"required":["createdAt","email","id","locale","plan","workspaceId"]},"ItemPageResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"}},"required":["items","page","size","totalElements"]},"QuotaInfo":{"type":"object","properties":{"plan":{"type":"string"},"limitMicros":{"type":"integer","format":"int64"},"usedMicros":{"type":"integer","format":"int64"},"exceeded":{"type":"boolean"}},"required":["exceeded","plan","usedMicros"]},"UsageBreakdownEntry":{"type":"object","properties":{"task":{"type":"string"},"costMicros":{"type":"integer","format":"int64"},"inputTokens":{"type":"integer","format":"int64"},"outputTokens":{"type":"integer","format":"int64"}},"required":["costMicros","inputTokens","outputTokens","task"]},"UsageSummaryResponse":{"type":"object","properties":{"from":{"type":"string","format":"date-time"},"to":{"type":"string","format":"date-time"},"totalInputTokens":{"type":"integer","format":"int64"},"totalOutputTokens":{"type":"integer","format":"int64"},"totalCostMicros":{"type":"integer","format":"int64"},"topCostDrivers":{"type":"array","items":{"$ref":"#/components/schemas/UsageBreakdownEntry"}},"quota":{"$ref":"#/components/schemas/QuotaInfo"}},"required":["from","quota","to","topCostDrivers","totalCostMicros","totalInputTokens","totalOutputTokens"]},"ProviderInfo":{"type":"object","properties":{"id":{"type":"string","enum":["OPENAI","ANTHROPIC","GOOGLE","AZURE","OPENROUTER","OLLAMA"]},"displayName":{"type":"string"},"requiresBaseUrl":{"type":"boolean"}},"required":["displayName","id","requiresBaseUrl"]}}}}